Revert blanket un-inline (broke >5-arg helpers); clean up build warnings
The un-inline change made helpers with more than 5 args real BPF-to-BPF calls,
which BPF forbids (no arg stack): 'stack arguments are not supported'. Restore
always_inline so it compiles. Also cast the two ctl.c mgmt-port loop bounds to
int (-Wsign-compare) and add -Wno-pass-failed to the BPF flags to silence the
cosmetic 'loop not unrolled' optimizer notes. Clean baseline; E2BIG verifier
limit is the next, separate problem.

+++ src/user/ctl.c
@@ -914,13 +914,13 @@ static void ctl_emit_config(const struct fw_config *c, struct sbuf *d)
 	sb_key(d, "dataplane_pref");     sb_u64(d, c->dataplane_pref);     sb_putc(d, ',');
 
 	sb_key(d, "mgmt_tcp_ports"); sb_putc(d, '[');
-	for (i = 0; i < (int)c->mgmt_tcp_count && i < CALY_MGMT_PORTS_MAX; i++) {
+	for (i = 0; i < (int)c->mgmt_tcp_count && i < (int)CALY_MGMT_PORTS_MAX; i++) {
 		if (i) sb_putc(d, ',');
 		sb_u64(d, c->mgmt_tcp_ports[i]);
 	}
 	sb_puts(d, "],");
 	sb_key(d, "mgmt_udp_ports"); sb_putc(d, '[');
-	for (i = 0; i < (int)c->mgmt_udp_count && i < CALY_MGMT_PORTS_MAX; i++) {
+	for (i = 0; i < (int)c->mgmt_udp_count && i < (int)CALY_MGMT_PORTS_MAX; i++) {
 		if (i) sb_putc(d, ',');
 		sb_u64(d, c->mgmt_udp_ports[i]);
 	}
